Add nl2bash.download
