+++ configs/nl2bash/nl2code_evaluate.yaml
@@ -9,8 +9,7 @@ device:
 output_dir: "output/output"
 
 dataset:
-  type: mlprogram.datasets.nl2bash.load
-  path: ./datasets/nl2bash.json
+  type: mlprogram.datasets.nl2bash.download
 
 encoder:
   word_encoder:
