+++ mlprogram/datasets/nl2bash/__init__.py
@@ -1,3 +1,3 @@
-from mlprogram.datasets.nl2bash.load import load  # noqa
+from mlprogram.datasets.nl2bash.download import download # noqa
 from mlprogram.datasets.nl2bash.functions import TokenizeQuery # noqa
 from mlprogram.datasets.nl2bash.functions import SplitValue  # noqa
